% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bereken_fishersom.R
\name{bereken_fishersom}
\alias{bereken_fishersom}
\title{bereken fishersom bedekking}
\usage{
bereken_fishersom(p)
}
\arguments{
\item{p}{bedekking in procent (numeric)}
}
\description{
Bij vegetatieanalyse, met name bij gelaagde vegetatie (bijvoorbeeld de
ondergroei onder een bladerdak), is de som van de bedekkingsgraden voor
verschillende soorten of lagen vaak hoger dan 100\% vanwege overlappende
begroeiing. Om dit probleem aan te pakken, heeft Fisher (2014) een methode
ontwikkeld om een gecombineerde "werkelijke" bedekkingswaarde te berekenen,
die ervoor zorgt dat deze niet hoger is dan 100\% en rekening houdt met de
verticale structuur.
}
