% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bereken_indicator.R
\name{bereken_indicator}
\alias{bereken_indicator}
\title{bereken indicatoren van een of meer opname(n)}
\usage{
bereken_indicator(
  naam_kopdata,
  naam_opnamen,
  dir_invoerdata,
  df_soorten = NULL,
  df_soortenlijstjes = NULL,
  df_indicatoren = NULL,
  df_habsleutels = NULL,
  herbereken_lagen = TRUE
)
}
\arguments{
\item{naam_kopdata}{String. De naam van de tabel met de kopgegevens
(metadata)
Dit is een csv-bestand.}

\item{naam_opnamen}{String. De naam van de tabel met de opnamen/relevés
Dit is een csv-bestand.}

\item{dir_invoerdata}{String. De naam van de map (inclusief pad) met de
gegevens van de gebruiker}

\item{df_soorten}{Dataframe. Soortenlijst, zie \code{maak_soortenlijst()}}

\item{df_soortenlijstjes}{Dataframe. Tabel (lang formaat) met de
standaardnaam van de soorten die tot een soortengroep gerekend worden, die
gebruikt wordt bij de determinatie en een uniek identificatienummer voor deze
soortengroep. Het is een systeembestand}

\item{df_indicatoren}{Dataframe. Overzichtstabel van de soortenlijstnummers
en met de opgave van een naam. Het is een systeembestand.}

\item{df_habsleutels}{Dataframe. Lijst van de (potentiële)
determineersleutels. Deze lijst bevat per vegetatiegroep (bijv. graslanden,
bossen) een lijstcode (1 karakter), een naam en een aanduiding of voor deze
vegetatiegroep er effectief al een digitale determineersleutel bestaat.
Het is een systeembestand.}

\item{herbereken_lagen}{Logical. Indicatoren die op vegetatielagen slaan
herberekenen (default = TRUE)?}
}
\value{
Een dataframe waarbij de kopgegevens aangevuld/bijgewerkt werden met
betrekking tot een aantal indicatoren
}
\description{
Deze functie berekent voor een of meer opname(n) | relevé(s) een aantal
standaard indicatorwaarden. Zie de details-sectie voor meer info.
}
\details{
We maken onderscheid tussen indicatoren die rechtstreeks ('on the fly')
tijdens de uitvoering van een SQL-opdracht worden berekend en indicatoren
die op voorhand worden berekend en die tijdens een SQL-opdracht opgehaald
kunnen worden. Het gaat hier over de tweede groep: we noemen ze de 'vaste'
indicatoren. De andere noemen we 'tijdelijke' indicatoren.
Vaste indicatoren zijn indicatoren die ofwel meermaals worden aangeroepen
ofwel indicatoren met een diagnostische waarde. Het onderscheid tussen een
tijdelijke en vaste indicator is wat subjectief.
In de tabel 'df_indicatoren' zijn de vaste indicatoren van de tijdelijke te
onderscheiden doordat de kolom 'veldnaam' niet leeg is.
Deze verwijst naar de kolom van de tabel met de kopgegevens waarin deze
indicatorwaarde moet bewaard worden.

Daarenboven maken we een onderscheid tussen indicatoren die een aantal
en die een som van bedekkingen uitdrukken. Dit wordt aangegeven in de kolom
aard.
}
