% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_en_voeg_kolommen_toe.R
\name{check_en_voeg_kolommen_toe}
\alias{check_en_voeg_kolommen_toe}
\title{helper functie om ontbrekende kolommen toe te voegen}
\usage{
check_en_voeg_kolommen_toe(df, benodigde_kolommen)
}
\arguments{
\item{df}{Dataframe}

\item{benodigde_kolommen}{Vector met de namen van kolommen die moeten
aanwezig zijn}
}
\description{
Deze functie inspecteert bij een dataframe of de opgegeven kolomnamen
aanwezig zijn.
Als ze ontbreken worden ze toegevoegd.
}
