% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_soortnamen.R
\name{check_soortnamen}
\alias{check_soortnamen}
\title{controleer soortnamen}
\usage{
check_soortnamen(
  custom_taxa,
  opname_soorten,
  bekende_namen_lower,
  officiele_taxa_lower,
  prefix_id = "CU",
  laatste_nummer = 0,
  file_custom
)
}
\arguments{
\item{custom_taxa}{dataframe met de soortnamen enkel bekend bij de gebruiker}

\item{opname_soorten}{vector van de soortnamen gebruikt in de tabel met de
opnamen}

\item{bekende_namen_lower}{vector van de soortnamen (in kleine letters)
die reeds bekend zijn.}

\item{officiele_taxa_lower}{vector van de soortnamen (in kleine letters),
beperkt tot de standaard inbo-lijst}

\item{prefix_id}{character de prefix die aan de soortnummers wordt toegevoegd}

\item{laatste_nummer}{numeric het huidige hoogste soortnummer van custom_taxa}

\item{file_custom}{character de bestandsnaam voor custom_taxa (inclusief pad)}
}
\description{
De functie vergelijkt de soortnamen van de opgegeven opnamen met de lijst
aanwezig in het package en met de standaard inbo-soortenlijst.
Ontbreekt een soortnaam in deze lijsten dan kan men deze toevoegen aan een
eigen soortenlijst.
Deze lijst wordt bewaard in een door de gebruiker opgegeven locatie.
}
