% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_update_sql.R
\name{doe_update_sql}
\alias{doe_update_sql}
\title{voer bijwerk-SQLregels geschreven voor SQL-Access uit in SQLite}
\usage{
doe_update_sql(con, sql, temp_tabelnaam = "tmp_update_logic")
}
\arguments{
\item{con}{Een \code{DBIConnection} object (SQLite-databank)}

\item{sql}{String. Een SQL-regel die een Access bijwerk-bewerking kan
uitvoeren.}

\item{temp_tabelnaam}{String. Een naam voor de tijdelijke tabel in de
SQLite-databank}
}
\value{
het resultaat van de bijwerkquery is een tabel onder de vorm van een
dataframe
}
\description{
Deze functie voert bijwerkqueries in een SQLite databank uit op
basis van bijwerkqueries ('Update ... set') voor Access.
}
\details{
Dit wordt in twee stappen uitgevoerd.
Eerst wordt van de updatequery omgezet in een create-query die een tijdelijke
tabel aanmaakt (met de soms complexe linken). Ze bevat een eenduidige unieke
link met de feitelijke doeltabel die moet bijgewerkt worden.
Het bijwerken van deze doeltabel wordt dan in tweede instantie uitgevoerd.
Doordat tussen de tijdelijke tabel en de doeltabel een eenduidige, simpele
link bestaat, is dit een vrij eenvoudige operatie, veel eenvoudiger en dus
betrouwbaarder dan de update-query willen vertalen in één keer.
}
\seealso{
Other converteer_sql: 
\code{\link{verwerk_sql_regels}()},
\code{\link{wijzig_select_into}()},
\code{\link{wijzig_veldnamen_acc}()},
\code{\link{zoekvervang_sql}()}
}
\concept{converteer_sql}
