% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sqlite_db.R
\name{import_sqlite_db}
\alias{import_sqlite_db}
\title{importeer basisdata naar een SQLite databank}
\usage{
import_sqlite_db(
  con,
  df_kop_indicator = NULL,
  df_opname = NULL,
  dir_invoerdata = NULL,
  df_soorten,
  df_soortenlijstjes,
  df_regels
)
}
\arguments{
\item{con}{Een \code{DBIConnection} object (SQLite-databank)}

\item{df_kop_indicator}{Dataframe. De tabel met de kopgegevens
met ook de berekende indicatorwaarden ervan. Default verwijst het naar een
dataframe dat het resultaat is van \code{bereken_indicator}}

\item{df_opname}{Dataframe. De tabel met opname(n)}

\item{dir_invoerdata}{String. De naam van de map (inclusief pad) met de
gegevens van de gebruiker}

\item{df_soorten}{Dataframe. Soortenlijst, zie \code{maak_soortenlijst()}}

\item{df_soortenlijstjes}{Dataframe. Tabel (lang formaat) met de
standaardnaam van de soorten die tot een soortengroep gerekend worden, die
gebruikt wordt bij de determinatie en een uniek identificatienummer voor deze
soortengroep. Het is een systeembestand}

\item{df_regels}{Dataframe. De tabel met de SQL-regels waarmee een opname kan
gekarakteriseerd worden.}
}
\description{
Deze functie importeert een reeks basisgegevens in een
SQLite-databank. Deze gegevens dienen om de opnamen te karakteriseren met
behulp van een set SQL-regels.
}
