% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karakteriseer_opname.R
\name{karakteriseer_opname}
\alias{karakteriseer_opname}
\title{karakteriseer een of meer opname(n)}
\usage{
karakteriseer_opname(
  naam_kopdata,
  naam_opnamen,
  dir_invoerdata,
  df_soorten = NULL,
  df_regels = NULL,
  df_soortenlijstjes = NULL,
  df_regel_selectie = file.path(system.file(package = "habkey"), "basisdata",
    "tbl_sleutel_uitkomst_hoofdtabel.csv"),
  sleutel_naam_sql = c("O", "G", "H", "M"),
  ...
)
}
\arguments{
\item{naam_kopdata}{String. De naam van de tabel met de kopgegevens
(metadata)
Dit is een csv-bestand.}

\item{naam_opnamen}{String. De naam van de tabel met de opnamen/relevés
Dit is een csv-bestand.}

\item{dir_invoerdata}{String. De naam van de map (inclusief pad) met de
gegevens van de gebruiker}

\item{df_soorten}{Dataframe. Soortenlijst, zie \code{maak_soortenlijst()}}

\item{df_regels}{Dataframe. De tabel met de SQL-regels waarmee een opname kan
gekarakteriseerd worden.}

\item{df_soortenlijstjes}{Dataframe. Tabel (lang formaat) met de
standaardnaam van de soorten die tot een soortengroep gerekend worden, die
gebruikt wordt bij de determinatie en een uniek identificatienummer voor deze
soortengroep. Het is een systeembestand}

\item{df_regel_selectie}{Dataframe. De tabel waarin alle kolommen die
beginnen met 'Regel' verwijzen naar een vraag/regel in de determineersleutel.
In deze functie wordt alleen beroep gedaan op de kolomnamen van de tabel.}

\item{sleutel_naam_sql}{String. Een karakter dat verwijst naar een specifieke
determineersleutel. Mogelijke waarden zijn c("O", "G", "H", "M"), die resp.
verwijzen naar de hoofd-, grasland-, heide- en moerassleutel.}

\item{...}{overige parameters, bijv. overschrijf = TRUE of herbereken = TRUE}
}
\value{
een dataframe waarin elke vraag een eigen kolom heeft en elke
opname een eigen regel
}
\description{
Deze functie typeert een vegetatieopname aan de hand van een
aantal kenmerken.
Deze kenmerken zijn de uitkomsten = antwoorden op een reeks vragen van een
determineersleutel.
}
\details{
De vragen zijn in eerste instantie omgezet in een reeks SQL-queries binnen
MS-Access. Deze sql-regels moeten hier ingevoerd worden.
}
