% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laad_kopdata.R
\name{laad_kopdata}
\alias{laad_kopdata}
\title{laad data met de kopgegevens in}
\usage{
laad_kopdata(naam_kopdata, dir_invoerdata)
}
\arguments{
\item{naam_kopdata}{String. De naam van de tabel met de kopgegevens
(metadata)
Dit is een csv-bestand.}

\item{dir_invoerdata}{String. De naam van de map (inclusief pad) met de
gegevens van de gebruiker}
}
\value{
Een bestand met de kopgegevens als dataframe
}
\description{
Deze functie controleert het opnamebestand en laadt ze in het werkgeheugen.
}
