% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laad_of_check_df.R
\name{laad_of_check_df}
\alias{laad_of_check_df}
\title{laad basis data}
\usage{
laad_of_check_df(df, bestandsnaam, package_naam = "habkey", custom_msg = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{bestandsnaam}{String. Naam van het systeembestand}

\item{package_naam}{String. Naam van het package met het systeembestand}

\item{custom_msg}{String. Foutboodschap}
}
\description{
Deze functie controleert eenvoudigweg wanneer de gebruiker eigen data wil
gebruiken in plaats van systeemdata of deze eigen data in een dataframe
zitten. Als geen bestandsnaam werd opgegeven dan wordt het systeembestand
ingeladen.
}
