% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laad_of_maak_soorten.R
\name{laad_of_maak_soorten}
\alias{laad_of_maak_soorten}
\title{laad soortenlijst of maak ze aan}
\usage{
laad_of_maak_soorten(datanaam = "soorten", ...)
}
\arguments{
\item{datanaam}{String. Naam van de soortenlijst. Als het een eigen
ingebrachte soortenlijst is dan moet het er een volledige verwijzing
naar zijn (inclusief pad en bestandsnaam en extensie), alleen csv-bestanden.}

\item{...}{Extra argumenten die direct worden doorgegeven aan
\code{maak_soortenlijst} (zoals \code{naam_opnamen}, \code{dir_invoerdata} en
\code{dir_achtergronddata}).}
}
\value{
Een soortenlijst als dataframe
}
\description{
Deze functie controleert of een soortenlijst specifiek voor de opnamen al
bestaat. Ze heeft
Als dat niet het geval is, wordt ze aangemaakt.
}
