% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laad_opnamen.R
\name{laad_opnamen}
\alias{laad_opnamen}
\title{laad opnamebestand in}
\usage{
laad_opnamen(naam_opnamen, dir_invoerdata)
}
\arguments{
\item{naam_opnamen}{String. De naam van de tabel met de opnamen/relevés
Dit is een csv-bestand.}

\item{dir_invoerdata}{String. De naam van de map (inclusief pad) met de
gegevens van de gebruiker}
}
\value{
Een opnamebestand als dataframe
}
\description{
Deze functie controleert het opnamebestand en laadt ze in het werkgeheugen.
}
