% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lees_csv2.R
\name{lees_csv2}
\alias{lees_csv2}
\title{lees een csv-bestand in}
\usage{
lees_csv2(map, bestand, foutboodschap = NULL, col_types = NULL)
}
\arguments{
\item{map}{De naam van de map waarin het bestand zit}

\item{bestand}{De naam van het csv-bestand}

\item{foutboodschap}{De boodschap die vermeld wordt,
als het bestand ontbreekt.}

\item{col_types}{One of \code{NULL}, a \code{\link[readr:cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, all column types will be inferred from \code{guess_max} rows of the
input, interspersed throughout the file. This is convenient (and fast),
but not robust. If the guessed types are wrong, you'll need to increase
\code{guess_max} or supply the correct types yourself.

Column specifications created by \code{\link[=list]{list()}} or \code{\link[readr:cols]{cols()}} must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[readr:cols_only]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
\itemize{
\item c = character
\item i = integer
\item n = number
\item d = double
\item l = logical
\item f = factor
\item D = date
\item T = date time
\item t = time
\item ? = guess
\item _ or - = skip
}

By default, reading a file without a column specification will print a
message showing what \code{readr} guessed they were. To remove this message,
set \code{show_col_types = FALSE} or set \code{options(readr.show_col_types = FALSE)}.}
}
\description{
Hulpfunctie om csv-bestanden met ';' als kolomscheidingsteken met
readr::read_csv2 in te lezen. Er wordt gecontroleerd of het bestand bestaat
en van het juiste type is.
}
