% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak_soortenlijst.R
\name{maak_soortenlijst}
\alias{maak_soortenlijst}
\title{maak een soortenlijst op basis van de opname(n)}
\usage{
maak_soortenlijst(
  naam_opnamen,
  dir_invoerdata,
  dir_achtergronddata = file.path(system.file(package = "habkey"), "basisdata")
)
}
\arguments{
\item{naam_opnamen}{String. De naam van de tabel met de opnamen/relevés
Dit is een csv-bestand.}

\item{dir_invoerdata}{String. De naam van de map (inclusief pad) met de
gegevens van de gebruiker}

\item{dir_achtergronddata}{String. De naam van de map met basisgegevens,
zoals standaardlijsten.
Standaardlijsten worden met het package meegeleverd.}
}
\description{
De functie stelt een soortenlijst samen op basis van de standaard flora-lijst
, een soortenlijst die specifiek voor dit package werd samengesteld en
(optioneel) een soortenlijst van de gebruiker.#'
De soortenlijst bevat alleen soortnamen die aanwezig zijn in het bestand met
de opname(n). Naast de standaardnaam bevat de tabel ook soortinformatie die
nuttig is bij het determineren van de opname(n).
}
\details{
Deze lijst wordt bewaard in een door de gebruiker opgegeven locatie.
}
