% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_taxonnaam.R
\name{parse_taxonnaam}
\alias{parse_taxonnaam}
\title{vereenvoudig de soortnaam door de auteurs te verwijderen}
\usage{
parse_taxonnaam(taxonnaam, parse_type = "canonicalnamewithmarker")
}
\arguments{
\item{taxonnaam}{String. Wetenschappelijke naam of namen die vereenvoudigd
moeten worden}

\item{parse_type}{String. Welk type vereenvoudiging teruggegeven moet worden.
Standaard is dit 'canonicalnamewithmarker', andere opties zijn
'canonicalname' en 'canonicalnamecomplete'}
}
\value{
Deze functie geeft de vereenvoudigde naam of namen terug (als string
of vector van strings)
}
\description{
Deze functie vereenvoudigt de opgegeven taxonnaam/taxonnamen
door de auteursnaam te verwijderen. Ze is overgenomen uit het package LSVI en
is gebaseerd op de functie name_parse uit het package rgbif.
}
