% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verwerk_sql_regels.R
\name{verwerk_sql_regels}
\alias{verwerk_sql_regels}
\title{Voer SQL regels iteratief uit}
\usage{
verwerk_sql_regels(
  con,
  sleutelvragen,
  df_regels,
  sleutel_naam_sql,
  naam_doeltabel,
  mapping_list
)
}
\arguments{
\item{con}{Een \code{DBIConnection} object (SQLite-databank)}

\item{sleutelvragen}{String. Een label waarmee eenduidig een link kan gelegd
worden met een vraag in de determinatiesleutel}

\item{df_regels}{Dataframe. De tabel met de SQL-regels waarmee een opname kan
gekarakteriseerd worden.}

\item{sleutel_naam_sql}{String. Een karakter dat verwijst naar een specifieke
determineersleutel. Mogelijke waarden zijn c("O", "G", "H", "M"), die resp.
verwijzen naar de hoofd-, grasland-, heide- en moerassleutel.}

\item{naam_doeltabel}{String. De naam voor de tabel (dataframe) waarin de
resultaten van de karakterisatie van een determinatiesleutel bewaard worden.}

\item{mapping_list}{Named vector. Een vector met veldnamen die
gebruikt worden in Access-SQL, maar die in R een andere naam krijgen:
"AccessNaam" = "R_Naam". In het package is het standaard een systeemvector,
zie \code{karakteriseer_opname}.}
}
\description{
Interne hulpfunctie om de geneste SQL-loops te verwerken.
}
\seealso{
Other converteer_sql: 
\code{\link{doe_update_sql}()},
\code{\link{wijzig_select_into}()},
\code{\link{wijzig_veldnamen_acc}()},
\code{\link{zoekvervang_sql}()}
}
\concept{converteer_sql}
