% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wijzig_select_into.R
\name{wijzig_select_into}
\alias{wijzig_select_into}
\title{converteer "Select ... into"-sql-regels naar SQLite "CREATE TABLE ..."
statements.}
\usage{
wijzig_select_into(con, sql)
}
\arguments{
\item{con}{Een \code{DBIConnection} object (SQLite-databank)}

\item{sql}{String. Een SQL-regel die 'INTO' bevat.}
}
\value{
een string met de gewijzigde SQL-regel
}
\description{
Deze functie wijzigt de syntax van queries die in Access
gebruikt worden om tabellen te maken in een syntax die hetzelfde doet in
SQLite. Indien de naam van de tabel al bekend is in de SQLite zal deze uit de
databank zonder waarschuwing verwijderd worden.
De conversie slaat alleen op SQL-regels die het INTO-statement bezitten.
Andere SQL-regels blijven onveranderd.
}
\seealso{
Other converteer_sql: 
\code{\link{doe_update_sql}()},
\code{\link{verwerk_sql_regels}()},
\code{\link{wijzig_veldnamen_acc}()},
\code{\link{zoekvervang_sql}()}
}
\concept{converteer_sql}
