% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wijzig_veldnamen_acc.R
\name{wijzig_veldnamen_acc}
\alias{wijzig_veldnamen_acc}
\title{wijzig veldnamen gebruikt in Access door namen SQLite}
\usage{
wijzig_veldnamen_acc(sql, mapping_list)
}
\arguments{
\item{sql}{String. Een SQL-regel}

\item{mapping_list}{Named vector. Een vector met veldnamen die
gebruikt worden in Access-SQL, maar die in R een andere naam krijgen:
"AccessNaam" = "R_Naam". In het package is het standaard een systeemvector,
zie \code{karakteriseer_opname}.}
}
\value{
een string met de gewijzigde SQL-regel
}
\description{
Deze functie wijzigt een aantal veldnamen die gebruikt worden
in de Access-databank, door de namen die in de SQLite-databank van toepassing
zijn.
}
\details{
De vervanging gebeurt op basis van een "named vector" van strings, waarbij de
namen van de strings verwijzen naar de nieuwe namen.
}
\seealso{
Other converteer_sql: 
\code{\link{doe_update_sql}()},
\code{\link{verwerk_sql_regels}()},
\code{\link{wijzig_select_into}()},
\code{\link{zoekvervang_sql}()}
}
\concept{converteer_sql}
