% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoekvervang_sql.R
\name{zoekvervang_sql}
\alias{zoekvervang_sql}
\title{zoek en vervang tekst in SQL}
\usage{
zoekvervang_sql(sql, naam_doeltabel, huidige_regel)
}
\arguments{
\item{sql}{String. Een SQL-regel}

\item{naam_doeltabel}{String. De naam van de tabel waarin de karakterisering
van de opname(n) wordt bewaard.}

\item{huidige_regel}{String. De kolomnaam waarin het antwoord op een
specifieke vraag van de determineersleutel wordt bewaard. Deze bestaat
standaard uit een regelnummer (verwijzend naar de determineersleutel) en een
label die toepasselijk is voor de vraag in de sleutel.}
}
\value{
een string met de gewijzigde SQL-tekst
}
\description{
Deze functie voert een aantal generieke zoek en vervangfuncties
uit in de SQL-regels
}
\details{
De SQLregels zijn opgemaakt in een Access-databank. De aanpassingen omvatten
twee soorten van generieke aanpassingen.
\enumerate{
\item De namen voor tabellen en velden verschillen daar met deze die in de
SQL-databank gebruikt worden.
\item De SQL-taal van MS-Access en SQLite zijn verschillend. Er zijn enkele
functies in Access die met een (vrij) eenvoudige zoek en vervang-methode
SQLite-compatibel kunnen gemaakt worden:
}
\itemize{
\item aanpassen van wildcarts (enkel *) door SQL-taal (\%)
\item aanpassen van de nz()-functie van Access door de SQL-functie COALESCE()
\item aanpassen van Iif()-functie
}
\enumerate{
\item De logische variabelen (TRUE/FALSE) worden door resp. 1 en 0 vervangen.
}
}
\seealso{
Other converteer_sql: 
\code{\link{doe_update_sql}()},
\code{\link{verwerk_sql_regels}()},
\code{\link{wijzig_select_into}()},
\code{\link{wijzig_veldnamen_acc}()}
}
\concept{converteer_sql}
